Переносит функции для отрисовки в utils/render.js
